+++ lib/src/__tests__/e2e/DatePicker.test.tsx
@@ -1,18 +1,21 @@
+import { ReactWrapper } from 'enzyme';
 import React from 'react';
+import DatePicker, { DatePickerProps } from '../../DatePicker/DatePicker';
 import { mount, utilsToUse } from '../test-utils';
-import DatePicker from '../../DatePicker/DatePicker';
 
 describe('e2e - DatePicker', () => {
-  let component;
+  let component: ReactWrapper<DatePickerProps>;
   const onChangeMock = jest.fn();
 
   beforeEach(() => {
     jest.clearAllMocks();
-    component = mount(<DatePicker
-      animateYearScrolling={false}
-      date={utilsToUse.date('2018-01-01T00:00:00.000Z')}
-      onChange={onChangeMock}
-    />);
+    component = mount(
+      <DatePicker
+        animateYearScrolling={false}
+        date={utilsToUse.date('2018-01-01T00:00:00.000Z')}
+        onChange={onChangeMock}
+      />
+    );
   });
 
   it('Should renders', () => {
@@ -24,16 +27,24 @@ describe('e2e - DatePicker', () => {
   });
 
   it('Should dispatch onChange on day click', () => {
-    component.find('Day button').at(2).simulate('click');
+    component
+      .find('Day button')
+      .at(2)
+      .simulate('click');
     expect(onChangeMock).toHaveBeenCalled();
   });
 
   it('Should render year selection', () => {
-    component.find('ToolbarButton').first().simulate('click');
+    component
+      .find('ToolbarButton')
+      .first()
+      .simulate('click');
     expect(component.find('Year').length).toBe(201);
 
-    component.find('Year').at(1).simulate('click');
+    component
+      .find('Year')
+      .at(1)
+      .simulate('click');
     expect(onChangeMock).toHaveBeenCalled();
   });
 });
-
